segcraft/requirements.txt
streamlit>=1.32.0
pydantic>=2.6.0
python-dotenv>=1.0.0
openai>=1.30.0
pandas>=2.1.0
tabulate>=0.9.0
Pillow>=10.0.0
